LayoutTests/http/tests/site-isolation/focus-navigation-cross-origin-iframe-expected.txt

Test that keyboard accessory bar next/previous navigation works for form elements inside a cross-origin iframe when site isolation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS First input focused in cross-origin iframe
PASS focusedElementId is "second"
PASS focusedElementIdAfterPrevious is "first"
PASS successfullyParsed is true

TEST COMPLETE
